+++ src/adapters/git.ts
@@ -38,7 +38,6 @@ abstract class GitAdapter implements IRepoAdapter {
       await this.git(repo).fetch('origin');
     } else {
       const git = simpleGit();
-      git.silent(true);
       await git.clone(repoPath, localPath, ['--depth', '1']);
     }
 
@@ -70,9 +69,7 @@ abstract class GitAdapter implements IRepoAdapter {
   }
 
   public async pushRepo(repo: IRepo, force: boolean): Promise<void> {
-    const options = {
-      '--force': force,
-    };
+    const options = force ? ['--force'] : undefined;
 
     await this.git(repo).push('origin', 'HEAD', options);
   }
@@ -94,9 +91,7 @@ abstract class GitAdapter implements IRepoAdapter {
   protected abstract getRepositoryUrl(repo: IRepo): string;
 
   protected git(repo: IRepo): SimpleGit {
-    const git = simpleGit(this.getRepoDir(repo));
-    git.silent(true);
-    return git;
+    return simpleGit(this.getRepoDir(repo));
   }
 
   protected isShepherdCommitMessage(message: string): boolean {